Condensed the color detection code into an easy-to-implement method, and added camera size standardization to Constants. Deleted unnecessary class.
