--- Team6220_CENTERSTAGE/src/main/java/org/firstinspires/ftc/team6220_CENTERSTAGE/ColorDetectionTests/DetectorFramework.java
@@ -1,6 +1,0 @@
-package org.firstinspires.ftc.team6220_CENTERSTAGE.ColorDetectionTests;
-
-import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-
-abstract public class DetectorFramework extends LinearOpMode {
-}
